requirements.txt
beautifulsoup4==4.3.2
fabric==1.8.0
gitpython==0.3.2rc1
hgapi==1.7.1
IPy==0.81
psycopg2==2.5.1
pytest==2.4.1
pyyaml==3.10
selenium==2.35.0
walkdir==0.3
